+++ src/pages/projects.tsx
@@ -1,0 +1,11 @@
+// src/pages/projects.tsx
+import MainLayout from '../components/layout/MainLayout';
+import ExplorePage from '../components/projects/ProjectsPage';
+
+export default function Projects() {
+  return (
+    <MainLayout title="Projects Hub - FUN(TIME)">
+      <ExplorePage />
+    </MainLayout>
+  );
+}